Fixed many bugs, ensure features working as intended.
Everything changed. Breaking changes everywhere. But getting closer to a stable place

+++ jobs/backupCore.go
@@ -4,18 +4,20 @@ import (
 	"errors"
 	"io"
 	"slices"
+	"strconv"
 	"time"
 
 	"github.com/ethanrous/weblens/fileTree"
 	"github.com/ethanrous/weblens/internal/log"
 	"github.com/ethanrous/weblens/internal/werror"
 	"github.com/ethanrous/weblens/models"
+	"github.com/ethanrous/weblens/models/rest"
 	"github.com/ethanrous/weblens/service/proxy"
 	"github.com/ethanrous/weblens/task"
 )
 
 func BackupD(interval time.Duration, pack *models.ServicePack) {
-	if pack.InstanceService.GetLocal().GetRole() != models.BackupServer {
+	if pack.InstanceService.GetLocal().GetRole() != models.BackupServerRole {
 		log.Error.Println("Backup service cannot be run on non-backup instance")
 		return
 	}
@@ -40,24 +42,16 @@ func BackupD(interval time.Duration, pack *models.ServicePack) {
 
 func BackupOne(core *models.Instance, pack *models.ServicePack) (*task.Task, error) {
 	meta := models.BackupMeta{
-		Core:                core,
-		FileService:         pack.FileService,
-		UserService:         pack.UserService,
-		WebsocketService:    pack.ClientService,
-		InstanceService:     pack.InstanceService,
-		TaskService:         pack.TaskService,
-		Caster:              pack.Caster,
-		AccessService:       pack.AccessService,
-		ProxyFileService:    &proxy.ProxyFileService{Core: core},
-		ProxyJournalService: &proxy.ProxyJournalService{Core: core},
-		ProxyUserService:    proxy.NewProxyUserService(core),
-		ProxyMediaService:   &proxy.ProxyMediaService{Core: core},
+		Core:             core,
+		FileService:      pack.FileService,
+		UserService:      pack.UserService,
+		WebsocketService: pack.ClientService,
+		InstanceService:  pack.InstanceService,
+		TaskService:      pack.TaskService,
+		Caster:           pack.Caster,
+		AccessService:    pack.AccessService,
 	}
 	return pack.TaskService.DispatchJob(models.BackupTask, meta, nil)
-}
-
-type serverInfoResponse struct {
-	Info models.ServerInfo `json:"info"`
 }
 
 func DoBackup(t *task.Task) {
@@ -65,34 +59,39 @@ func DoBackup(t *task.Task) {
 
 	t.OnResult(
 		func(r task.TaskResult) {
-			meta.Caster.PushTaskUpdate(t, "backup_progress", r)
+			meta.Caster.PushTaskUpdate(t, models.BackupProgressEvent, r)
 		},
 	)
 
 	t.SetErrorCleanup(
 		func(errTsk *task.Task) {
+			err := errTsk.ReadError()
+			if err == nil {
+				log.Error.Println("Trying to show error in backup task, but error is nil")
+				return
+			}
+
 			meta.Caster.PushTaskUpdate(
-				errTsk, "backup_failed",
-				task.TaskResult{"coreId": meta.Core.ServerId(), "error": errTsk.ReadError().Error()},
+				errTsk, models.BackupFailedEvent,
+				task.TaskResult{"coreId": meta.Core.ServerId(), "error": err.Error()},
 			)
 		},
 	)
 
 	localRole := meta.InstanceService.GetLocal().GetRole()
-	if localRole == models.InitServer {
+	if localRole == models.InitServerRole {
 		t.ReqNoErr(werror.ErrServerNotInitialized)
-	} else if localRole != models.BackupServer {
+	} else if localRole != models.BackupServerRole {
 		t.ReqNoErr(werror.ErrServerIsBackup)
 	}
 
-	coreClient := meta.WebsocketService.GetClientByServerId(meta.Core.ServerId())
-	if coreClient == nil {
-		t.ReqNoErr(werror.Errorf("Core websocket not connected"))
-		// Dead code
-		return
-	}
+	// Get the active websocket client for the core
+	// coreClient := meta.WebsocketService.GetClientByServerId(meta.Core.ServerId())
+	// if coreClient == nil {
+	// 	t.Fail(werror.Errorf("Core websocket not connected"))
+	// }
 
-	log.Debug.Printf("Starting backup of [%s]", coreClient.GetRemote().GetName())
+	// log.Debug.Printf("Starting backup of [%s]", coreClient.GetRemote().GetName())
 
 	stages := models.NewBackupTaskStages()
 
@@ -100,89 +99,17 @@ func DoBackup(t *task.Task) {
 	t.SetResult(task.TaskResult{"stages": stages, "coreId": meta.Core.ServerId()})
 
 	// Read core server info and check if it is really a core server
-	req := proxy.NewCoreRequest(meta.Core, "GET", "").OverwriteEndpoint("/api/info")
-	infoRes, err := proxy.CallHomeStruct[serverInfoResponse](req)
+	req := proxy.NewCoreRequest(meta.Core, "GET", "/info")
+	infoRes, err := proxy.CallHomeStruct[rest.ServerInfo](req)
 	if err != nil {
 		t.ReqNoErr(err)
 	}
 
-	meta.Core.SetReportedRole(infoRes.Info.Role)
-	if infoRes.Info.Role != models.CoreServer {
-		t.ReqNoErr(werror.Errorf("Remote role is [%s] expected core", infoRes.Info.Role))
+	meta.Core.SetReportedRole(infoRes.Role)
+	if infoRes.Role != models.CoreServerRole {
+		t.ReqNoErr(werror.Errorf("Remote role is [%s] expected core", infoRes.Role))
 	}
 
-	stages.FinishStage("connecting")
-	stages.StartStage("fetching_users")
-	t.SetResult(task.TaskResult{"stages": stages})
-	// Backup users
-	users, err := meta.ProxyUserService.GetAll()
-	if err != nil {
-		t.ReqNoErr(err)
-	}
-
-	stages.FinishStage("fetching_users")
-	stages.StartStage("writing_users")
-	t.SetResult(task.TaskResult{"stages": stages})
-	for user := range users {
-		err = meta.UserService.Add(user)
-		if err != nil {
-			t.ReqNoErr(err)
-		}
-	}
-
-	stages.FinishStage("writing_users")
-	stages.StartStage("fetching_keys")
-	t.SetResult(task.TaskResult{"stages": stages})
-	// Fetch keys from core
-	keysRq := proxy.NewCoreRequest(meta.Core, "GET", "/backup/keys")
-	keys, err := proxy.CallHomeStruct[[]models.ApiKeyInfo](keysRq)
-	if err != nil {
-		t.ReqNoErr(err)
-	}
-
-	stages.FinishStage("fetching_keys")
-	stages.StartStage("writing_keys")
-	t.SetResult(task.TaskResult{"stages": stages})
-	// Add keys to access service
-	for _, key := range keys {
-		if _, err := meta.AccessService.GetApiKey(key.Key); err == nil {
-			continue
-		}
-
-		err = meta.AccessService.AddApiKey(key)
-		if err != nil {
-			t.ReqNoErr(err)
-		}
-	}
-
-	stages.FinishStage("writing_keys")
-	stages.StartStage("fetching_instances")
-	t.SetResult(task.TaskResult{"stages": stages})
-	// Fetch instances from core
-	instancesRq := proxy.NewCoreRequest(meta.Core, "GET", "/backup/instances")
-	instances, err := proxy.CallHomeStruct[[]*models.Instance](instancesRq)
-	if err != nil {
-		t.ReqNoErr(err)
-	}
-
-	stages.FinishStage("fetching_instances")
-	stages.StartStage("writing_instances")
-	t.SetResult(task.TaskResult{"stages": stages})
-	// Add instances to access service
-	for _, r := range instances {
-		if err := meta.InstanceService.Get(r.ServerId()); err == nil {
-			continue
-		}
-
-		err = meta.InstanceService.Add(r)
-		if err != nil {
-			t.ReqNoErr(err)
-		}
-	}
-
-	stages.FinishStage("writing_instances")
-	stages.StartStage("sync_journal")
-	t.SetResult(task.TaskResult{"stages": stages})
 	// Find most recent action timestamp
 	latest, err := meta.FileService.GetJournalByTree(meta.Core.Id).GetLatestAction()
 	if err != nil {
@@ -194,27 +121,95 @@ func DoBackup(t *task.Task) {
 		latestTime = latest.GetTimestamp()
 	}
 
-	log.Trace.Printf("Backup latest action is %s", latestTime.String())
+	log.Trace.Func(func(l log.Logger) { l.Printf("Backup latest action is %s", latestTime.String()) })
 
-	// Get new history updates
-	updatedLifetimes, err := meta.ProxyJournalService.GetLifetimesSince(latestTime)
+	stages.StartStage("fetching_backup_data")
+	t.SetResult(task.TaskResult{"stages": stages})
+
+	backupRq := proxy.NewCoreRequest(meta.Core, "GET", "/servers/backup").WithQuery("timestamp", strconv.FormatInt(latestTime.UnixMilli(), 10))
+	backupResponse, err := proxy.CallHomeStruct[rest.BackupInfo](backupRq)
 	t.ReqNoErr(err)
 
-	log.Trace.Printf("Backup got %d updated lifetimes from core", len(updatedLifetimes))
+	stages.StartStage("writing_users")
+	t.SetResult(task.TaskResult{"stages": stages})
 
-	stages.FinishStage("sync_journal")
+	log.Debug.Println(backupResponse.Users)
+
+	// Write the users to the users service
+	for _, userInfo := range backupResponse.Users {
+		user := rest.UserInfoArchiveToUser(userInfo)
+		err = meta.UserService.Add(user)
+		if err != nil {
+			t.ReqNoErr(err)
+		}
+	}
+
+	stages.StartStage("writing_keys")
+	t.SetResult(task.TaskResult{"stages": stages})
+
+	// Write keys to access service
+	for _, key := range backupResponse.ApiKeys {
+		if _, err := meta.AccessService.GetApiKey(key.Key); err == nil {
+			continue
+		}
+
+		err = meta.AccessService.AddApiKey(key)
+		if err != nil {
+			t.ReqNoErr(err)
+		}
+	}
+
+	stages.StartStage("writing_instances")
+	t.SetResult(task.TaskResult{"stages": stages})
+
+	// Write instances to access service
+	for _, serverInfo := range backupResponse.Instances {
+		instance := rest.ServerInfoToInstance(serverInfo)
+		if err := meta.InstanceService.Get(instance.ServerId()); err == nil {
+			continue
+		}
+
+		err = meta.InstanceService.Add(instance)
+		if err != nil {
+			t.ReqNoErr(err)
+		}
+	}
+
+	stages.StartStage("sync_journal")
+	t.SetResult(task.TaskResult{"stages": stages})
+
+	log.Trace.Func(func(l log.Logger) {
+		l.Printf("Backup got %d updated lifetimes from core", len(backupResponse.FileHistory))
+	})
+
 	stages.StartStage("sync_fs")
 	t.SetResult(task.TaskResult{"stages": stages})
 
 	// Sort lifetimes so that files created or moved most recently are updated last.
-	slices.SortFunc(updatedLifetimes, fileTree.LifetimeSorter)
+	slices.SortFunc(backupResponse.FileHistory, fileTree.LifetimeSorter)
 
 	journal := meta.FileService.GetJournalByTree(meta.Core.Id)
-	if len(updatedLifetimes) > 0 {
-		for _, lt := range updatedLifetimes {
+	if len(backupResponse.FileHistory) > 0 {
+		for _, lt := range backupResponse.FileHistory {
 			// Add the lifetime to the journal, or update it if it already exists
 			err = journal.Add(lt)
 			t.ReqNoErr(err)
+		}
+	}
+
+	if len(journal.GetAllLifetimes()) < backupResponse.LifetimesCount {
+		log.Debug.Func(func(l log.Logger) {
+			l.Printf("Backup journal is missing %d lifetimes", backupResponse.LifetimesCount-len(journal.GetAllLifetimes()))
+		})
+
+		req := proxy.NewCoreRequest(meta.Core, "GET", "/journal").WithQuery("timestamp", "0")
+		allLts, err := proxy.CallHomeStruct[[]*fileTree.Lifetime](req)
+		t.ReqNoErr(err)
+		for _, lt := range allLts {
+			if journal.Get(lt.ID()) == nil {
+				err = journal.Add(lt)
+				t.ReqNoErr(err)
+			}
 		}
 	}
 
@@ -224,17 +219,26 @@ func DoBackup(t *task.Task) {
 
 	pool := t.GetTaskPool().GetWorkerPool().NewTaskPool(true, t)
 	t.SetChildTaskPool(pool)
-	// meta.WebsocketService.TaskSubToPool(t.TaskId(), pool.GetRootPool().ID())
 
 	// Sort lifetimes so that files created or moved most recently are updated last.
 	// This is to make sure parent directories are created before their children
 	slices.SortFunc(activeLts, fileTree.LifetimeSorter)
 
+	// Check if the file already exists on the server and copy it if it doesn't
 	for _, lt := range activeLts {
 		latestMove := lt.GetLatestMove()
 
 		existingFile, err := meta.FileService.GetFileByTree(lt.ID(), meta.Core.ServerId())
-		if err == nil {
+		log.Debug.Printf("File %s exists: %v", lt.GetLatestPath(), existingFile != nil)
+
+		// If the file already exists, but is the wrong size, an earlier copy most likely failed. Delete it and copy it again.
+		if existingFile != nil && !existingFile.IsDir() && existingFile.Size() != lt.Actions[0].Size {
+			err = meta.FileService.DeleteFiles([]*fileTree.WeblensFileImpl{existingFile}, meta.Core.ServerId(), meta.Caster)
+			t.ReqNoErr(err)
+			existingFile = nil
+		}
+
+		if err == nil && existingFile != nil {
 			if latestMove.ActionType == fileTree.FileDelete {
 				err = meta.FileService.DeleteFiles(
 					[]*fileTree.WeblensFileImpl{existingFile}, meta.Core.ServerId(), meta.Caster,
@@ -252,7 +256,8 @@ func DoBackup(t *task.Task) {
 				t.ReqNoErr(err)
 			}
 			continue
-		} else if !errors.Is(err, werror.ErrNoFile) {
+
+		} else if err != nil && !errors.Is(err, werror.ErrNoFile) {
 			t.Fail(err)
 		}
 
@@ -261,8 +266,8 @@ func DoBackup(t *task.Task) {
 				continue
 			}
 
-			_, err := meta.FileService.GetFileByContentId(lt.ContentId)
-			if err == nil {
+			f, err := meta.FileService.GetFileByContentId(lt.ContentId)
+			if err == nil && f.Size() == lt.Actions[0].Size {
 				continue
 			} else if !errors.Is(err, werror.ErrNoFile) {
 				t.Fail(err)
@@ -278,17 +283,16 @@ func DoBackup(t *task.Task) {
 			continue
 		}
 
-		if !coreClient.IsOpen() {
-			coreClient = meta.WebsocketService.GetClientByServerId(meta.Core.ServerId())
-		}
+		// if !coreClient.IsOpen() {
+		// 	coreClient = meta.WebsocketService.GetClientByServerId(meta.Core.ServerId())
+		// }
 
 		copyFileMeta := models.BackupCoreFileMeta{
-			ProxyFileService: meta.ProxyFileService,
-			FileService:      meta.FileService,
-			File:             restoreFile,
-			Caster:           meta.Caster,
-			Core:             meta.Core,
-			Filename:         filename,
+			FileService: meta.FileService,
+			File:        restoreFile,
+			Caster:      meta.Caster,
+			Core:        meta.Core,
+			Filename:    filename,
 		}
 
 		_, err = meta.TaskService.DispatchJob(models.CopyFileFromCoreTask, copyFileMeta, pool)
@@ -325,7 +329,7 @@ func DoBackup(t *task.Task) {
 func CopyFileFromCore(t *task.Task) {
 	meta := t.GetMeta().(models.BackupCoreFileMeta)
 	t.SetErrorCleanup(func(t *task.Task) {
-		meta.Caster.PushTaskUpdate(t, "copy_file_failed", task.TaskResult{"filename": meta.Filename, "coreId": meta.Core.ServerId()})
+		meta.Caster.PushTaskUpdate(t, models.CopyFileFailedEvent, task.TaskResult{"filename": meta.Filename, "coreId": meta.Core.ServerId()})
 	})
 
 	filename := meta.Filename
@@ -334,11 +338,11 @@ func CopyFileFromCore(t *task.Task) {
 	}
 
 	meta.Caster.PushPoolUpdate(
-		t.GetTaskPool(), "copy_file_started", task.TaskResult{
+		t.GetTaskPool(), models.CopyFileStartedEvent, task.TaskResult{
 			"filename": filename, "coreId": meta.Core.ServerId(), "timestamp": time.Now().UnixMilli(),
 		},
 	)
-	log.Trace.Printf("Copying file from core [%s]", meta.File.Filename())
+	log.Trace.Func(func(l log.Logger) { l.Printf("Copying file from core [%s]", meta.File.Filename()) })
 
 	if meta.File.GetContentId() == "" {
 		t.ReqNoErr(werror.WithStack(werror.ErrNoContentId))
@@ -385,7 +389,7 @@ func RestoreCore(t *task.Task) {
 	type restoreInitParams struct {
 		Name     string            `json:"name"`
 		Role     models.ServerRole `json:"role"`
-		Key      models.ApiKeyInfo `json:"usingKeyInfo"`
+		Key      models.ApiKey     `json:"usingKeyInfo"`
 		RemoteId string            `json:"remoteId"`
 		LocalId  string            `json:"localId"`
 	}
@@ -394,32 +398,32 @@ func RestoreCore(t *task.Task) {
 	t.SetErrorCleanup(
 		func(errTsk *task.Task) {
 			meta.Pack.Caster.PushTaskUpdate(
-				errTsk, "restore_failed", task.TaskResult{"error": errTsk.ReadError().Error()},
+				errTsk, models.RestoreFailedEvent, task.TaskResult{"error": errTsk.ReadError().Error()},
 			)
 		},
 	)
 
 	// Prime server to be restored. This will fail if the server is already initialized
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Connecting to remote", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Connecting to remote", "timestamp": time.Now().UnixMilli()},
 	)
 
 	key, err := meta.Pack.AccessService.GetApiKey(meta.Core.UsingKey)
 	t.ReqNoErr(err)
 
 	initParams := restoreInitParams{
-		Name: meta.Core.Name, Role: models.RestoreServer, Key: key, RemoteId: meta.Local.Id,
+		Name: meta.Core.Name, Role: models.RestoreServerRole, Key: key, RemoteId: meta.Local.Id,
 		LocalId: meta.Core.Id,
 	}
 
-	_, err = proxy.NewCoreRequest(meta.Core, "POST", "").OverwriteEndpoint("/api/init").WithBody(initParams).Call()
+	_, err = proxy.NewCoreRequest(meta.Core, "POST", "/servers/init").WithBody(initParams).Call()
 	if err != nil {
 		t.ReqNoErr(err)
 	}
 
 	// Restore journal
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Restoring file history", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Restoring file history", "timestamp": time.Now().UnixMilli()},
 	)
 	lts := meta.Pack.FileService.GetJournalByTree(meta.Core.Id).GetAllLifetimes()
 	_, err = proxy.NewCoreRequest(meta.Core, "POST", "/restore/history").WithBody(lts).Call()
@@ -427,26 +431,22 @@ func RestoreCore(t *task.Task) {
 
 	// Restore users
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Restoring users", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Restoring users", "timestamp": time.Now().UnixMilli()},
 	)
 
 	usersIter, err := meta.Pack.UserService.GetAll()
 	t.ReqNoErr(err)
 
-	var users []map[string]any
+	var users []rest.UserInfo
 	for u := range usersIter {
-		formatted, err := u.FormatArchive()
-		if err != nil {
-			t.ReqNoErr(err)
-		}
-		users = append(users, formatted)
+		users = append(users, rest.UserToUserInfo(u))
 	}
 	_, err = proxy.NewCoreRequest(meta.Core, "POST", "/restore/users").WithBody(users).Call()
 	t.ReqNoErr(err)
 
 	// Restore keys
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Restoring api keys", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Restoring api keys", "timestamp": time.Now().UnixMilli()},
 	)
 	rootUser := meta.Pack.UserService.GetRootUser()
 	keys, err := meta.Pack.AccessService.GetAllKeysByServer(rootUser, meta.Core.ServerId())
@@ -457,7 +457,7 @@ func RestoreCore(t *task.Task) {
 
 	// Restore instances
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Restoring api keys", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Restoring api keys", "timestamp": time.Now().UnixMilli()},
 	)
 
 	instances := meta.Pack.InstanceService.GetAllByOriginServer(meta.Core.ServerId())
@@ -466,7 +466,7 @@ func RestoreCore(t *task.Task) {
 
 	// Restore files
 	meta.Pack.Caster.PushTaskUpdate(
-		t, "restore_progress", task.TaskResult{"stage": "Restoring files", "timestamp": time.Now().UnixMilli()},
+		t, models.RestoreProgressEvent, task.TaskResult{"stage": "Restoring files", "timestamp": time.Now().UnixMilli()},
 	)
 	for i, lt := range lts {
 		latest := lt.GetLatestAction()
@@ -500,7 +500,7 @@ func RestoreCore(t *task.Task) {
 		}
 
 		meta.Pack.Caster.PushTaskUpdate(
-			t, "restore_progress", task.TaskResult{"files_total": len(lts), "files_restored": i},
+			t, models.RestoreProgressEvent, task.TaskResult{"filesTotal": len(lts), "filesRestored": i},
 		)
 	}
 
@@ -509,8 +509,8 @@ func RestoreCore(t *task.Task) {
 		t.ReqNoErr(err)
 	}
 
-	meta.Core.SetReportedRole(models.CoreServer)
-	meta.Pack.Caster.PushTaskUpdate(t, "restore_complete", nil)
+	meta.Core.SetReportedRole(models.CoreServerRole)
+	meta.Pack.Caster.PushTaskUpdate(t, models.RestoreCompleteEvent, nil)
 
 	// Disconnect the core client to force a reconnection
 	coreClient := meta.Pack.ClientService.GetClientByServerId(meta.Core.ServerId())
